requirements.txt
black
flake8
ipython
isort
jupyterlab
loguru
matplotlib
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
tqdm
typer
langchain
openai
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
psycopg2
pinecone-client
pgvector
zhipuai
sentence_transformers
langchain_community
bilibili-api-python
langchainhub 
httpx_sse 
langchain_openai 
langchain_huggingface
gradio
modelscope